configure checkpoint file path on cmd line.
