+++ src/components/abciapp/src/abci/server/callback/mod.rs
@@ -7,7 +7,7 @@ mod utils;
 use {
     crate::{
         abci::{
-            config::global_cfg::CFG, server::ABCISubmissionServer, staking, IN_SAFE_ITV,
+            server::ABCISubmissionServer, staking, IN_SAFE_ITV,
             POOL,
         },
         api::{
@@ -21,11 +21,12 @@ use {
         ResponseBeginBlock, ResponseCheckTx, ResponseCommit, ResponseDeliverTx,
         ResponseEndBlock, ResponseInfo, ResponseInitChain, ResponseQuery,
     },
+    config::abci::global_cfg::CFG,
     fp_storage::hash::{Sha256, StorageHasher},
     lazy_static::lazy_static,
     ledger::{
         converter::is_convert_account,
-        staking::{KEEP_HIST, CHECKPOINT},
+        staking::{KEEP_HIST},
         store::{
             api_cache,
             fbnc::{new_mapx, Mapx},
@@ -80,7 +81,7 @@ pub fn info(s: &mut ABCISubmissionServer, req: &RequestInfo) -> ResponseInfo {
     TENDERMINT_BLOCK_HEIGHT.swap(h, Ordering::Relaxed);
     resp.set_last_block_height(h);
     if 0 < h {
-        if CHECKPOINT.disable_evm_block_height < h && h < CHECKPOINT.enable_frc20_height
+        if CFG.checkpoint.disable_evm_block_height < h && h < CFG.checkpoint.enable_frc20_height
         {
             resp.set_last_block_app_hash(la_hash);
         } else {
@@ -141,8 +142,8 @@ pub fn check_tx(s: &mut ABCISubmissionServer, req: &RequestCheckTx) -> ResponseC
             resp
         }
         TxCatalog::EvmTx => {
-            if CHECKPOINT.disable_evm_block_height < td_height
-                && td_height < CHECKPOINT.enable_frc20_height
+            if CFG.checkpoint.disable_evm_block_height < td_height
+                && td_height < CFG.checkpoint.enable_frc20_height
             {
                 resp.code = 2;
                 resp.log = "EVM is disabled".to_owned();
@@ -205,8 +206,8 @@ pub fn begin_block(
         pnk!(la.update_staking_simulator());
     }
 
-    if CHECKPOINT.disable_evm_block_height < header.height
-        && header.height < CHECKPOINT.enable_frc20_height
+    if CFG.checkpoint.disable_evm_block_height < header.height
+        && header.height < CFG.checkpoint.enable_frc20_height
     {
         ResponseBeginBlock::default()
     } else {
@@ -249,8 +250,8 @@ pub fn deliver_tx(
                         }
                     }
 
-                    if CHECKPOINT.disable_evm_block_height < td_height
-                        && td_height < CHECKPOINT.enable_frc20_height
+                    if CFG.checkpoint.disable_evm_block_height < td_height
+                        && td_height < CFG.checkpoint.enable_frc20_height
                     {
                         if is_convert_account(&tx) {
                             resp.code = 2;
@@ -316,8 +317,8 @@ pub fn deliver_tx(
             resp
         }
         TxCatalog::EvmTx => {
-            if CHECKPOINT.disable_evm_block_height < td_height
-                && td_height < CHECKPOINT.enable_frc20_height
+            if CFG.checkpoint.disable_evm_block_height < td_height
+                && td_height < CFG.checkpoint.enable_frc20_height
             {
                 resp.code = 2;
                 resp.log = "EVM is disabled".to_owned();
@@ -386,8 +387,8 @@ pub fn end_block(
         &begin_block_req.byzantine_validators.as_slice(),
     );
 
-    if td_height <= CHECKPOINT.disable_evm_block_height
-        || td_height >= CHECKPOINT.enable_frc20_height
+    if td_height <= CFG.checkpoint.disable_evm_block_height
+        || td_height >= CFG.checkpoint.enable_frc20_height
     {
         let _ = s.account_base_app.write().end_block(req);
     }
@@ -416,8 +417,8 @@ pub fn commit(s: &mut ABCISubmissionServer, req: &RequestCommit) -> ResponseComm
     let la_hash = state.get_state_commitment().0.as_ref().to_vec();
     let cs_hash = s.account_base_app.write().commit(req).data;
 
-    if CHECKPOINT.disable_evm_block_height < td_height
-        && td_height < CHECKPOINT.enable_frc20_height
+    if CFG.checkpoint.disable_evm_block_height < td_height
+        && td_height < CFG.checkpoint.enable_frc20_height
     {
         r.set_data(la_hash);
     } else {
