+++ src/ledger/Cargo.toml
@@ -25,8 +25,9 @@ unicode-normalization = "0.1.13"
 time = "0.2.26"
 tendermint = "0.19.0"
 indexmap = { version = "1.6.2", features = ["serde"] }
+config = { path = "../components/config" }
 fp-types = { path = "../components/contracts/primitives/types" }
-toml = "0.5.8"
+
 ruc = "1.0"
 zei = { git = "https://github.com/FindoraNetwork/zei", tag = "v0.1.4x" }
 zeiutils = { package = "utils", git = "https://github.com/FindoraNetwork/zei", tag = "v0.1.4x" }
